requirements.txt
Django==4.2.3
django-ckeditor==6.6.1
Pillow==10.0.0
django-richtextfield==1.6.1